backend/requirements.txt
fastapi
uvicorn

sqlalchemy 
psycopg2-binary

python-jose[cryptography]
passlib[bcrypt]
python-multipart

bcrypt==4.0.1

python-dotenv